+++ src/components/base/Text/Text.types.ts
@@ -1,0 +1,8 @@
+export interface TextProps {
+  children: React.ReactNode;
+  color?: string;
+  weight?: "normal" | "bold";
+  fontWeight?: number;
+  fontSize?: string;
+  fontFamily?: string;
+}