+++ src/components/base/index.ts
@@ -1,5 +1,5 @@
 export { default as Button } from "./Button";
 export { default as Tag } from "./Tag";
-export { default as Text } from "./Text";
-export { default as Paragraph } from "./Paragraph";
+export { default as Text } from "./Text/Text";
+export { default as Paragraph } from "./Text/Paragraph";
 export { default as Div } from "./Div";